Circ Chain Functions should take BIGINTs
Because INTEGER isn't suitable for circ IDs.

This only requires editing the input definitions.

+++ Open-ILS/src/sql/Pg/090.schema.action.sql
@@ -784,7 +784,7 @@ CREATE TYPE action.circ_chain_summary AS (
 );
 
 
-CREATE OR REPLACE FUNCTION action.circ_chain ( ctx_circ_id INTEGER ) RETURNS SETOF action.circulation AS $$
+CREATE OR REPLACE FUNCTION action.circ_chain ( ctx_circ_id BIGINT ) RETURNS SETOF action.circulation AS $$
 DECLARE
     tmp_circ action.circulation%ROWTYPE;
     circ_0 action.circulation%ROWTYPE;
@@ -819,7 +819,7 @@ BEGIN
 END;
 $$ LANGUAGE 'plpgsql';
 
-CREATE OR REPLACE FUNCTION action.summarize_circ_chain ( ctx_circ_id INTEGER ) RETURNS action.circ_chain_summary AS $$
+CREATE OR REPLACE FUNCTION action.summarize_circ_chain ( ctx_circ_id BIGINT ) RETURNS action.circ_chain_summary AS $$
 
 DECLARE
